Add files via upload
This is a quiz website using HTML,CSS and JavaScript. Through this website users can select a quiz from a list of available options, which may include categories such as geography, science and mathematics.

Once a quiz is selected, users will be presented with a series of multiple-choice questions, and they will need to select the correct answer from a list of options. The application will keep track of their progress, and provide a score at the end of the quiz.


In addition to the basic functionality of taking quizzes, this quiz applications also include features such as progress tracking and leaderboard.